chore: rename sidebar component files
