+++ client/src/app/_components/global/platform-sidebar/LogoutButton.tsx
