+++ client/src/app/_components/global/platform-sidebar/SidebarItem.tsx
