+++ client/src/layouts/platform/index.tsx
@@ -1,7 +1,7 @@
 import { useEffect, useState } from 'react';
 import { usePathname } from 'next/navigation';
 
-import { PlatformSidebar } from '~/app/_components/global/platform-sidebar';
+import { PlatformSidebar } from '~/app/_components/global/platform-sidebar/PlatformSidebar';
 import { UserStoreStatus } from '~/app/_components/temp-user-store-status';
 import Providers from '~/providers';
 import RootLayout from '..';
